+++ src/test/scala/com/guizmaii/distances/DistanceApiSpec.scala
@@ -2,9 +2,9 @@ package com.guizmaii.distances
 
 import cats.effect.{Async, IO}
 import cats.temp.par.Par
+import com.guizmaii.distances.GoogleDistanceProvider.GoogleGeoApiContext
 import com.guizmaii.distances.Types.TravelMode.{Bicycling, Driving}
 import com.guizmaii.distances.Types.{Distance, LatLong, PostalCode, _}
-import com.guizmaii.distances.utils.GoogleGeoApiContext
 import monix.eval.Task
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{BeforeAndAfterEach, Matchers, WordSpec}
